move post save to releaseprojectenvironment
